backend/ai_service/requirements.txt
fastapi
uvicorn
gunicorn
redis
python-jose[cryptography]
pytest
flask-cors
